requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.28
openai>=1.14
python-dotenv>=1.0 